#241 refactor: Party 패키지 구조 변경
