+++ src/main/java/com/drinkeg/drinkeg/converter/CommentConverter.java
@@ -2,7 +2,7 @@ package com.drinkeg.drinkeg.converter;
 
 import com.drinkeg.drinkeg.domain.Comment;
 import com.drinkeg.drinkeg.domain.Member;
-import com.drinkeg.drinkeg.domain.Party;
+import com.drinkeg.drinkeg.party.domain.Party;
 import com.drinkeg.drinkeg.dto.CommentDTO.CommentRequestDTO;
 import com.drinkeg.drinkeg.dto.CommentDTO.CommentResponseDTO;
 import org.springframework.stereotype.Component;
