+++ src/main/java/com/drinkeg/drinkeg/converter/PartyBookmarkConverter.java
@@ -1,7 +1,7 @@
 package com.drinkeg.drinkeg.converter;
 
 import com.drinkeg.drinkeg.domain.Member;
-import com.drinkeg.drinkeg.domain.Party;
+import com.drinkeg.drinkeg.party.domain.Party;
 import com.drinkeg.drinkeg.domain.PartyBookmark;
 import org.springframework.stereotype.Component;
