+++ src/main/java/com/drinkeg/drinkeg/converter/PartyConverter.java
@@ -1,9 +1,9 @@
 package com.drinkeg.drinkeg.converter;
 
 import com.drinkeg.drinkeg.domain.Member;
-import com.drinkeg.drinkeg.domain.Party;
-import com.drinkeg.drinkeg.dto.PartyDTO.PartyRequestDTO;
-import com.drinkeg.drinkeg.dto.PartyDTO.PartyResponseDTO;
+import com.drinkeg.drinkeg.party.domain.Party;
+import com.drinkeg.drinkeg.party.dto.PartyRequestDTO;
+import com.drinkeg.drinkeg.party.dto.PartyResponseDTO;
 import org.springframework.stereotype.Component;
