+++ src/main/java/com/drinkeg/drinkeg/converter/PartyJoinMemberConverter.java
@@ -1,11 +1,9 @@
 package com.drinkeg.drinkeg.converter;
 
 import com.drinkeg.drinkeg.domain.Member;
-import com.drinkeg.drinkeg.domain.Party;
+import com.drinkeg.drinkeg.party.domain.Party;
 import com.drinkeg.drinkeg.domain.PartyJoinMember;
 import org.springframework.stereotype.Component;
-
-import java.time.LocalDateTime;
 
 @Component
 public class PartyJoinMemberConverter {
