+++ src/main/java/com/drinkeg/drinkeg/domain/PartyBookmark.java
@@ -1,5 +1,6 @@
 package com.drinkeg.drinkeg.domain;
 
+import com.drinkeg.drinkeg.party.domain.Party;
 import jakarta.persistence.*;
 import lombok.*;
