+++ src/main/java/com/drinkeg/drinkeg/party/service/PartyService.java
@@ -1,10 +1,8 @@
-package com.drinkeg.drinkeg.service.partyService;
+package com.drinkeg.drinkeg.party.service;
 
-import com.drinkeg.drinkeg.domain.Member;
-import com.drinkeg.drinkeg.domain.Party;
-import com.drinkeg.drinkeg.domain.PartyJoinMember;
-import com.drinkeg.drinkeg.dto.PartyDTO.PartyRequestDTO;
-import com.drinkeg.drinkeg.dto.PartyDTO.PartyResponseDTO;
+import com.drinkeg.drinkeg.party.domain.Party;
+import com.drinkeg.drinkeg.party.dto.PartyRequestDTO;
+import com.drinkeg.drinkeg.party.dto.PartyResponseDTO;
 import com.drinkeg.drinkeg.dto.loginDTO.commonDTO.PrincipalDetail;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
