+++ src/main/java/com/drinkeg/drinkeg/repository/PartyJoinMemberRepository.java
@@ -1,7 +1,7 @@
 package com.drinkeg.drinkeg.repository;
 
 import com.drinkeg.drinkeg.domain.Member;
-import com.drinkeg.drinkeg.domain.Party;
+import com.drinkeg.drinkeg.party.domain.Party;
 import com.drinkeg.drinkeg.domain.PartyJoinMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
