+++ src/main/java/com/drinkeg/drinkeg/service/commentService/CommentServiceImpl.java
@@ -5,7 +5,7 @@ import com.drinkeg.drinkeg.converter.CommentConverter;
 import com.drinkeg.drinkeg.converter.RecommentConverter;
 import com.drinkeg.drinkeg.domain.Comment;
 import com.drinkeg.drinkeg.domain.Member;
-import com.drinkeg.drinkeg.domain.Party;
+import com.drinkeg.drinkeg.party.domain.Party;
 import com.drinkeg.drinkeg.domain.Recomment;
 import com.drinkeg.drinkeg.dto.CommentDTO.CommentRequestDTO;
 import com.drinkeg.drinkeg.dto.CommentDTO.CommentResponseDTO;
@@ -14,7 +14,7 @@ import com.drinkeg.drinkeg.dto.loginDTO.commonDTO.PrincipalDetail;
 import com.drinkeg.drinkeg.exception.GeneralException;
 import com.drinkeg.drinkeg.repository.CommentRepository;
 import com.drinkeg.drinkeg.service.memberService.MemberService;
-import com.drinkeg.drinkeg.service.partyService.PartyService;
+import com.drinkeg.drinkeg.party.service.PartyService;
 import com.drinkeg.drinkeg.service.recommentService.RecommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
