+++ src/main/java/com/drinkeg/drinkeg/service/partyBookmarkService/PartyBookmarkServiceImpl.java
@@ -4,16 +4,16 @@ import com.drinkeg.drinkeg.apipayLoad.code.status.ErrorStatus;
 import com.drinkeg.drinkeg.converter.PartyBookmarkConverter;
 import com.drinkeg.drinkeg.converter.PartyConverter;
 import com.drinkeg.drinkeg.domain.Member;
-import com.drinkeg.drinkeg.domain.Party;
+import com.drinkeg.drinkeg.party.domain.Party;
 import com.drinkeg.drinkeg.domain.PartyBookmark;
-import com.drinkeg.drinkeg.dto.PartyDTO.PartyResponseDTO;
+import com.drinkeg.drinkeg.party.dto.PartyResponseDTO;
 import com.drinkeg.drinkeg.dto.loginDTO.commonDTO.PrincipalDetail;
 import com.drinkeg.drinkeg.event.partyBookmarkEvent.PartyBookmarkCreateEvent;
 import com.drinkeg.drinkeg.event.partyBookmarkEvent.PartyBookmarkDeleteEvent;
 import com.drinkeg.drinkeg.exception.GeneralException;
 import com.drinkeg.drinkeg.repository.PartyBookmarkRepository;
 import com.drinkeg.drinkeg.service.memberService.MemberService;
-import com.drinkeg.drinkeg.service.partyService.PartyService;
+import com.drinkeg.drinkeg.party.service.PartyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
